+++ changelogs/adapter/admob/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to Yandex Adapter for AdMob Mediation will be documented in this file.
 
+## Version 8.2.0.0
+
+#### Updated
+* Yandex Mobile Ads SDK version 8.2.0
+
 ## Version 8.1.0.0
 
 #### Updated
@@ -11,6 +16,11 @@ All notable changes to Yandex Adapter for AdMob Mediation will be documented in 
 
 #### Updated
 * Yandex Mobile Ads SDK version 8.0.0
+
+## Version 7.18.7.0
+
+#### Updated
+* Yandex Mobile Ads SDK version 7.18.7
 
 ## Version 7.18.6.0