+++ changelogs/mediation/applovin/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to AppLovin Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 13.5.1.2
+
+#### Updated
+* Yandex Mobile Ads SDK version 8.2.0
+
 ## Version 13.5.1.1
 
 #### Updated