+++ changelogs/mediation/ironsource/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to IronSource Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 9.2.0.2
+
+#### Updated
+* Yandex Mobile Ads SDK version 8.2.0
+
 ## Version 9.2.0.1
 
 #### Updated